+++ bootstrap.sh
@@ -1,7 +1,18 @@
-#!/bin/sh
+#!/bin/bash
 
-# see https://community.plone.org/t/not-using-bootstrap-py-as-default/620
-rm -r ./lib ./include ./local ./bin
-virtualenv --clear .
-./bin/pip install --upgrade pip setuptools zc.buildout
-./bin/buildout 
+if [ -x "$(which python)" ]; then
+    rm -r py2
+
+    virtualenv --clear --no-site-packages -p python py2
+
+    ./py2/bin/pip install coverage
+    ./py2/bin/pip install -e .[test]
+fi
+if [ -x "$(which python3)" ]; then
+    rm -r py3
+
+    virtualenv --clear --no-site-packages -p python3 py3
+
+    ./py3/bin/pip install coverage
+    ./py3/bin/pip install -e .[test]
+fi