+++ CHANGES.rst
@@ -5,7 +5,9 @@ History
 2.3.2 (unreleased)
 ------------------
 
-- No changes yet.
+- Fix ``yafowil.utils.cssid`` to return unicode in order to prevent malformed
+  rendering in python 3.
+  [rnix, 2019-03-25]
 
 
 2.3.1 (2018-11-07)